Refactor: Todo 구조 정리
