+++ lib/screens/todo.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mung_ge_mung_ge/models/todo.dart';
+import 'package:mung_ge_mung_ge/screens/components/todo-list.dart';
 
 class TodoScreen extends StatefulWidget {
   @override
@@ -144,109 +145,4 @@ class _TodoScreenState extends State<TodoScreen> {
   }
 }
 
-class TodoList extends StatelessWidget {
-  final List<Todo> todos;
-  final void Function(int index) clickCheckbox;
-  final void Function(int index) deleteTodo;
-  final void Function(int index) clickEditBtn;
-  final void Function(int oldIndex, int newIndex) onReorderTodoList;
-
-  TodoList({
-    required this.todos,
-    required this.clickCheckbox,
-    required this.deleteTodo,
-    required this.clickEditBtn,
-    required this.onReorderTodoList,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: Padding(
-        padding: EdgeInsets.symmetric(vertical: 10),
-        child: ReorderableListView(
-          onReorder: this.onReorderTodoList,
-          children: <Widget>[
-            for (int index = 0; index < todos.length; index++)
-              TodoItem(
-                key: ValueKey(todos[index]),
-                todo: todos[index],
-                clickCheckbox: () => clickCheckbox(index),
-                deleteTodo: () => deleteTodo(index),
-                clickEditBtn: () => clickEditBtn(index),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-
-class TodoItem extends StatelessWidget {
-  Todo todo;
-  VoidCallback clickCheckbox;
-  VoidCallback deleteTodo;
-  VoidCallback clickEditBtn;
-  
-  TodoItem({
-    Key? key,
-    required this.todo,
-    required this.clickCheckbox,
-    required this.deleteTodo,
-    required this.clickEditBtn,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Dismissible(
-      key: UniqueKey(),
-      onDismissed: (direction) {
-        deleteTodo();
-      },
-      child: ListTile(
-        key: UniqueKey(),
-        title: Container(
-          padding: EdgeInsets.only(
-            top: 10.0,
-            bottom: 10.0,
-            left: 5.0,
-            right: 20.0
-          ),
-          decoration: BoxDecoration(
-            border: Border(
-              bottom: BorderSide(
-                color: Colors.black.withOpacity(0.5),
-              ),
-            ),
-          ),
-          child: Row(
-            children: <Widget>[
-              Checkbox(
-                value: todo.isDone,
-                activeColor: Colors.grey,
-                onChanged: (event) => clickCheckbox(),
-              ),
-              Text(
-                todo.content,
-                style: TextStyle(
-                  fontSize: 16.0,
-                  decoration: todo.isDone ? TextDecoration.lineThrough : TextDecoration.none,
-                ),
-              ),
-              Spacer(),
-              GestureDetector(
-                child: Icon(
-                  Icons.edit,
-                  color: Theme.of(context).primaryColor,
-                ),
-                onTap: clickEditBtn,
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
